agents/requirements.txt
# gRPC dependencies
grpcio==1.70.0
grpcio-tools==1.70.0
grpcio-reflection==1.70.0

# LLM Providers SDKs
openai==1.61.1
google-genai==1.10.0
anthropic==0.45.2

# Utilities
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.7.1
httpx==0.28.1

# Logging
loguru==0.7.3
